requirements.txt
# Core dependencies for The AGI Assistant
# Python 3.10+ required

# Observation Layer
mss==9.0.1
sounddevice==0.4.6
numpy==1.24.3
pynput==1.7.6
psutil==5.9.5

# Processing Layer
faster-whisper==0.10.0
pytesseract==0.3.10
Pillow==10.0.0
opencv-python==4.8.0

# Intelligence Layer
ollama==0.1.6
sentence-transformers==2.2.2
scikit-learn==1.3.0

# Automation Layer
pyautogui==0.9.54
playwright==1.40.0
pygetwindow==0.0.9

# UI Layer
customtkinter==5.2.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2
tqdm==4.66.1

# Database
# SQLite is included in Python standard library

# Packaging (for creating .exe)
pyinstaller==6.3.0

